login aligned and bg added
my brain isn't braining
