+++ src/app/(auth)/login/page.tsx
@@ -1,94 +1,51 @@
-import Image from "next/image";
+"use client";
 import Link from "next/link";
 import { UserAuthForm } from "./user-auth-form";
-import { Command } from "lucide-react";
-import IconCloud from "@/components/ui/magicui/icon-cloud";
-
-const slugs = [
-  "typescript",
-  "javascript",
-  "dart",
-  "java",
-  "react",
-  "flutter",
-  "android",
-  "html5",
-  "css3",
-  "nodedotjs",
-  "express",
-  "nextdotjs",
-  "prisma",
-  "amazonaws",
-  "postgresql",
-  "firebase",
-  "nginx",
-  "vercel",
-  "testinglibrary",
-  "jest",
-  "cypress",
-  "docker",
-  "git",
-  "jira",
-  "github",
-  "gitlab",
-  "visualstudiocode",
-  "androidstudio",
-  "sonarqube",
-  "figma",
-];
+import { useTheme } from "next-themes";
 
 export default function AuthenticationPage() {
-  return (
-    <div className="container relative h-[800px] flex-col items-center justify-center md:grid lg:max-w-none lg:grid-cols-2 lg:px-0">
-      <div className="relative h-full flex-col bg-muted p-10 text-white lg:flex dark:border-r">
-        <div className="absolute inset-0 bg-zinc-900" />
-        <div className="relative z-20 flex items-center text-lg font-medium">
-          <Command size={24} className="mr-2" />
-          Teconiq-dev
-        </div>
-        <div className="relative z-20 mt-auto">
-          <div className="w-full flex flex-col justify-center">
-            <IconCloud iconSlugs={slugs} />
+  const { theme } = useTheme();
 
-            <blockquote className="space-y-2">
-              <p className="text-lg">
-                &ldquo;Education is the most powerful weapon which you can use
-                to change the world.&rdquo;
-              </p>
-              <footer className="text-sm">Nelson Mandela</footer>
-            </blockquote>
+  return (
+    <div className="mx-auto flex flex-1 w-full flex-col justify-center space-y-6 bg-transparent">
+      {
+        // light and dark custom background
+        theme === "light" ? (
+          <div className="absolute inset-0 z-[-2] h-full bg-white bg-[linear-gradient(to_right,#f0f0f0_1px,transparent_1px),linear-gradient(to_bottom,#f0f0f0_1px,transparent_1px)] bg-[size:6rem_4rem]">
+            <div className="absolute bottom-0 left-0 right-0 top-0 bg-[radial-gradient(circle_800px_at_100%_200px,#d5c5ff,transparent)]"></div>
           </div>
-        </div>
-      </div>
-      <div className="lg:p-8">
-        <div className="mx-auto flex w-full flex-col justify-center space-y-6 sm:w-[350px]">
-          <div className="flex flex-col space-y-2 text-center">
-            <h1 className="text-2xl font-semibold tracking-tight">
-              Students Login
-            </h1>
-            <p className="text-sm text-muted-foreground">
-              Enter your StudentID below to login
-            </p>
-          </div>
-          <UserAuthForm />
-          <p className="px-8 text-center text-sm text-muted-foreground">
-            By clicking continue, you agree to our{" "}
-            <Link
-              href="/terms"
-              className="underline underline-offset-4 hover:text-primary"
-            >
-              Terms of Service
-            </Link>{" "}
-            and{" "}
-            <Link
-              href="/privacy"
-              className="underline underline-offset-4 hover:text-primary"
-            >
-              Privacy Policy
-            </Link>
-            .
+        ) : (
+          <div className="absolute top-0 inset-0 z-[0] h-full bg-purple-950/10 bg-[radial-gradient(ellipse_20%_80%_at_50%_-20%,rgba(120,119,198,0.3),rgba(255,255,255,0))]"></div>
+        )
+      }
+
+      <div className="flex flex-col m-14 p-10 space-y-2 text-center transform-gpu dark:[border:1px_solid_rgba(255,255,255,.1)] dark:[box-shadow:0_-20px_80px_-20px_#8686f01f_inset]">
+        <div>
+          <h1 className="text-2xl font-semibold tracking-tight">
+            Students Login
+          </h1>
+          <p className="text-sm text-muted-foreground">
+            Enter your StudentID below to login
           </p>
         </div>
+        <UserAuthForm />
+        <p className="px-8 text-center text-sm text-muted-foreground">
+          By clicking continue, you agree to our{" "}
+          <Link
+            href="/terms"
+            className="underline underline-offset-4 hover:text-primary"
+          >
+            Terms of Service
+          </Link>{" "}
+          and{" "}
+          <Link
+            href="/privacy"
+            className="underline underline-offset-4 hover:text-primary"
+          >
+            Privacy Policy
+          </Link>
+          .
+        </p>
       </div>
     </div>
   );
