+++ src/components/ui/magicui/icon-cloud.tsx
@@ -17,7 +17,6 @@ export const cloudProps: Omit<ICloud, "children"> = {
       justifyContent: "center",
       alignItems: "center",
       width: "100%",
-      paddingTop: 40,
     },
   },
   options: {
